+++ test/app.e2e-spec.ts
@@ -3,6 +3,8 @@ import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
 import { App } from 'supertest/types';
 import { AppModule } from './../src/app.module';
+import { FileLoggerService } from '../src/logging/file-logger.service';
+import { HttpExceptionFilter } from '../src/logging/http-exception.filter';
 
 describe('AppController (e2e)', () => {
   let app: INestApplication<App>;
@@ -23,6 +25,7 @@ describe('AppController (e2e)', () => {
     }).compile();
 
     app = moduleFixture.createNestApplication();
+    app.useGlobalFilters(new HttpExceptionFilter(app.get(FileLoggerService)));
     await app.init();
 
     const setupStatus = await request(app.getHttpServer())